backend/requirements.txt
# ------------------------------

# 🚀 FastAPI Backend

# ------------------------------

fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.22

# ------------------------------

# 🔐 Auth & Security

# ------------------------------

PyJWT==2.9.0
passlib[bcrypt]==1.7.4
httpx==0.28.1
google-auth==2.47.0

# ------------------------------

# 🗄️ Database

# ------------------------------

pymongo==4.16.0
dnspython==2.8.0

# ------------------------------

# ⚙️ Config & Environment

# ------------------------------

python-dotenv==1.2.1
pydantic==2.12.5

# ------------------------------

# 📊 Your Existing Project Dependencies

# ------------------------------

altair==6.0.0
annotated-types==0.7.0
anyio==4.12.1
asgiref==3.11.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.5
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.4
cycler==0.12.1
distro==1.9.0
Django==6.0.2
dnspython==2.8.0
et_xmlfile==2.0.0
Flask==3.1.3
Flask-Login==0.6.3
fonttools==4.61.1
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.15
google-api-core==2.25.2
google-api-python-client==2.188.0
google-auth==2.47.0
google-auth-httplib2==0.3.0
google-genai==1.59.0
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.1
httpx==0.28.1
idna==3.11
image==1.5.33
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
narwhals==2.15.0
numpy==2.4.1
openpyxl==3.1.5
packaging==26.0
pandas==2.3.3
pandas-stubs==2.3.3.260113
pdfminer.six==20251230
pdfplumber==0.11.9
pillow==12.1.0
plotly==6.5.2
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==23.0.0
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pymongo==4.16.0
pyparsing==3.3.1
pypdf==6.7.0
PyPDF2==3.0.1
pypdfium2==5.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
qrcode==8.2
referencing==0.37.0
reportlab==4.4.10
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sqlparse==0.5.5
streamlit==1.53.1
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
watchdog==6.0.0
websockets==15.0.1
Werkzeug==3.1.6

# ------------------------------

# 🔑 (Future - Passkeys)

# ------------------------------

webauthn==1.11.0
